fix(apollo-parser): create nodes for all value kinds
We were only creating nodes for some of the Value kinds, like ObjectValue and ListValue, but were missing nodes for StringValue, FloatValue, IntValue, BooleanValue and NullValue. This commit introduces new nodes for these variants of the Value. We are also dropping creating a Value node, since by creating NullValue et al nodes, it is already considered to be a Value Node.

This commit fixes being able to iterate and query values in arguments and other nodes they are present.