follow up to #105: don't always pop tokens when creating errors
We should only sometimes pop the token off the error vector when creating errors.

For example, when a Name is missing in the Enum Type Definition, we should not be removing the current token, as the current token allows us to continue parsing the rest of the input correctly.

We still pop off the token when an incorrect definition is supplied, as we want to continue parsing other definitions.