docs: provide examples on using miette and annotate_snippets with apollo-parser (#118)
As the bulk of the diagnostics work is still to be done as part of the semantic analysis story, I am hesitant to prematurely add a Display implementation to errors. After all, we are not sure what sort of context will be useful from a semantics and validation point of view! Also, because apollo-parser is meant to be used in different environments for different use cases, it's yet to be determined whether as ascii Display would be beneficial to everyone.

Instead, this commit documents two different ways to display parsing errors that apollo-parser returns: using miette and annotate_snippets. For the moment, it'll be up to the implementors to determine their visual error output.

As part of this upgraded to 2021 edition.