fix(apollo-parser): add ignored tokens to TYPE nodes in correct place (#153)
This commit re-writes the parsing logic for TYPE nodes to have an easier time accounting for ignored tokens and their correct place in the AST. An intermediate representation of the TYPE and it's nullability and list status is created and then processed to put the correct tokens in their spots.

Before this commit this sort of query
```graphql
mutation MyMutation($custId: Int!, $b: String) {
  myMutation(custId: $custId)
}
```
would result the `ast.document.to_string()` to have this output:
```graphql
mutation MyMutation($custId: , Int!$b:  String) {
    myMutation(custId: $custId)
}
```
which is incorrect. The `to_string()` now results in the exact same output, as the AST created is correct.