apollo-parser: update operation definition tests after #158
Avoids creating an operation definition node when an unrelated token is popped off in the parsing of operation definition nodes.

Additionally, we are adding an correctly testing for number of errors and definitions in the AST when a description precedes an Operation Definition. We are also adding an error test case to the test suite with a description before an operation definition.