feat(apollo-compiler): create diagnostics with source spans (#251)
This PR introduces diagnostics to `apollo-compiler`. Given SyntaxNode information we have from `apollo-parser`, we are able to store that information in compiler's `Values` and create proper source spans. 

We are using [`miette`](https://docs.rs/miette/latest/miette/) to create diagnostics. Graphical display that `miette` provides is only available with `Display`. This means diagnostics can be used in myriads of other use cases that might not need the graphical interface (e.g. wasm). 

(almost) all `values` in `apollo-compiler` store a pointer back to the AST which can be accessed with `ast_node()` getter methods.