feat(apollo-compiler): add relationship between field types and definitions they might be referring to
Let's say we have a graphql schema that looks like this:

```
type Person {
  name: String
  picture(size: Number): Url
}

enum Number {
    INT
    FLOAT
}

scalar Url @specifiedBy(url: "https://tools.ietf.org/html/rfc3986")
```

We want to be able to easily access the `picture` field's Url definition and its attributes, as well as the `picture` argument's definition (`Number`). This establishes a relationship between `Type` and various definitions, as well as provides a few easier to use access methods.

All GraphQL definitions are not also grouped under `Definition` value in the database. In the `db`, they can be accessed via `db_definitions()`. The separation between `definitions` and `db_definitions` ensures we don't prefill the entire database in one go, and still proceed to set up queries as they are called.