feat(apollo-compiler): add field types to operation definitions (#257)
We want to be able to have access to type information in an executable definitions, i.e. tie schema information to query information. This PR introduces some basic relationships for more straightforward selection sets. A given `Field` now has both the `reference_ty_id` and `ty` information when available. This makes it possible to have access to schema-defined types. 

Co-authored-by: Irina Shestak <8107784+lrlna@users.noreply.github.com>