fix(apollo-compiler): do not add fragment_id to fragment_spread
because fragment definitions can have fragment spreads, we are running into a self-referential cycle when searching for a fragment definition to get its id. Instead, search for the fragment definition when getting a fragment in a fragment spread in the wrapper API.

Adds tests for nested fragments, and an introspection query, where this bug has initially arisen.

closes #281