feat(compiler): thread-safe compiler db (#295)
closes #275 

To make sure we can implement `salsa::ParallelDatabase` for the compiler's `RootDatabase`, we needed to remove any `!Send` return values in any of the `query_group` queries. The `SyntaxTree` returned by `db.ast()` query is at the root (get it?) of pretty much all queries, and was the first to be fixed. This PR makes a slight change to `apollo-parser`'s `SyntaxTree`. It now stores `GreenNode`s directly, and only creates a typed tree when asked (`SyntaxTree::document()`). This is actually nicer, and is non-breaking change to `apollo-parser`, as we only add a `fn syntax_node()` to the API surface. 

For the moment, this PR also removes `db.definitions()`, which means that for the moment we remove some of the incrementality of the query groups. The next step in this process is to establish a direct relationship between compiler's HIR to Ast Nodes, which should mean we can bring back this query.