refactor(compiler): create a validation database (#303)
Validation was its own module, which isn't as useful as having it be a salsa query group! We can now take advantage of memoised validation functions, as well as have a more succinct RootDatabase.

In this process, we've also renamed all query groups to match their associated storage names, as well as add `Database` to their naming. This ensures there is no confusion to `Document` or `Definition` being confused with ast's document or definition for the code reader. 

To allow `ValidationDatabase` to continue making calls to specific queries in `DocumentDatabase`, we've also added an `Upcast` trait.