chore(parser): reduce token copying (#323)
* Reduce token copying

Since the original lexer results are not needed anymore after this step,
we can take ownership of the tokens and errors vectors and reverse them
in-place without making a copy. Big schemas can have 100K+ tokens so
it's actually quite a lot of work to copy them.

* Reduce double-clones of tokens in the parser

Some of these clones were not necessary. In particular the `.expect`
method cloned the token unconditionally (including the string inside)
and then cloned the string again immediately afterwards. This removes
the first clone by reordering the `current.index()` call to satisfy the
borrow checker.

The `.data().to_string()` clone is only used in the error case, but
avoiding that will require more work.