feat(compiler): pub compiler storage (#328)
* feat(compiler): export database storage in a more accessible way

While trying to come up with a nice example to show how the compiler's query_groups can be used in other contexts, I realised we are not exporting any of our internal storage structs. This allows for internal query_groups to be exported, and adds a specific example writing a linter.

It was a bit difficult to create lints with `name` fields in all of the hir nodes not having their own ast_nodes attached to them, so this also adds ast node pointers to all the names. 

This allows to build on apollo-compiler's storage in separate databases. For example, it's now possible to

```rust
use apollo_compiler::{database::{AstStorage, DocumentStorage}};

#[salsa::database(AstStorage, DoumentStorage)]
pub struct AnotherDatabase {
    pub storage: salsa::Storage<AnotherDatabase>,
}
```

* feat(compiler): represent Name as its own HIR node

This allows us to have an ast_node attached to every name node, which means we can point to it exactly in diagnostics. Adds a `name_src` to Definitions impl.

* chore(compiler): add an example of extending various compiler storage

* chore(compiler): update `Into` impl for Name to `From`

* chore(compiler): name hir node