AST int → HIR float  when expected by typesystem (best effort)
The syntax for an integer value is also valid for a float value,
but the latter has a wider range of accepted values (~53 bits v.s. 32 bits).
This adds support for such "extra" values during HIR creation
by returning a `hir::Value::Float` for an `ast::IntValue`
when type system definitions indicate that a float is expected.

In some cases, type system definitions would have enough information
but this does not yet implement all lookups necessary.

In other cases it truly does not have sufficient information.
Either way, `hir::Value::Int` is used as a fallback as before.