use a single input type
This changes the inputs system a bit to use a single salsa input.
Salsa panics if a query can't be resolved. For inputs, that means it
will panic if you request an input which doesn't have a value. So if we
have separate input queries for GraphQL schemas, queries, and documents,
we have to know beforehand which type an input is to request its source
code or file path or whatever. That's technically possible with the
source manifest but it seems like more work.

With this change the input type is a _structure_ that has a SourceType
(one of schema/query/document) and a source string. The source string is
`Arc<str>` to make it cheaper to clone. We can also use `Arc<str>` in
miette Diagnostics (technically I did it because I "have to", but it's
also beneficial for memory use). This requires miette 5 to work, though,
so I updated it :zany_face: