feat(compiler): validate directive locations + undefined directives (#417)
partial of #246

This looks over each directive usage and checks whether it's in a correct location as specified by its original DirectiveDefinition. In case of incorrect usage `UnsupportedLocation` error is returned. Since we are already looking for the original DirectiveDefinition, this PR also adds a `UndefinedDefinition` error for undefined directives in case that DirectiveDefinition is not found.

Co-authored-by: Renée <renee@kooi.me>