feat(compiler): add meta types to compiler context  (#489)
closes #439 

Introduces introspection types, built in scalars and built in directives in a form of a source file to the compiler. This allows us to properly validate operations with introspection fields, as well as provide more detailed diagnostics when working with other built in types. 

This PR introduces a new source type - `BuiltIn`. `SourceType` now also has appropriate `is_{ty}` methods:
- `is_built_in`
- `is_document`
- `is_query`, and
- `is_schema`.

There are also a few changes to the `HirDatabase` in type queries. Implicit and built-in types should not be returned as part of the regular queries to the compiler context, so we added an additional queries to separate between user provided types and built in types:
- `db.object_types` returns only user supplied object types
- `db.object_types_with_built_ins` returns user supplied and built in (introspection) object types
- `db.enums` returns only user supplied enum types
- `db.enums_with_built_ins` returns user supplied and built in (introspection) enum types

Finding field by name on particular type does search through implicit fields (`__type`, `__typename` and `__schema`). So `type_definition.field(&compiler.db, "__type")` will return the implicit field's field definition.

---------

Co-authored-by: Irina Shestak <8107784+lrlna@users.noreply.github.com>