feat(compiler): validate fragments are used, validate fragment type condition (#483)
Implements *fragment spread types existence* validation: "fragments must be specified on types that exist in the schema. This applies for both named and inline fragments. If they are not defined in the schema, the fragment is invalid."

Implements *fragments on composite types* validation: "Fragments can only be declared on unions, interfaces, and objects."

Validates fragments are used. 

`ValidationDatabase` has 2 new methods:
- `validate_fragment_type_condition` which validates that a type condition is valid: it exisits in the schema, and is on a composite type
- validate_fragment_used` which validates that a fragment is used in a query.
---------

Co-authored-by: Camille Lawrence <camillelawrence@Camilles-MacBook-Pro.local>
Co-authored-by: Iryna Shestak <shestak.irina@gmail.com>