feat(parser): add SyntaxTree::token_limit(), similar to recursion_limit()
This enables finding out how many tokens were present in a succesful parse,
which can be useful to choose where to set the limit.

Also change the default token limit from no limit to usize::MAX.
This is almost always the same in terms of whether the limit
will be reached (especially on 64-bit architecture) but makes the tracker
always be present even when not explicitly configured,
allowing the new method to always return a meaningful value.