feat(compiler): add as_$type() methods to hir::Value
We usually expected people to just `match` on the HIR values, but that's
a little verbose if you just need one type. For example the router uses
this custom function:

```rust
fn as_string(value: &hir::Value) -> Option<&String> {
    if let hir::Value::String(string) = value {
        Some(string)
    } else {
        None
    }
}
```

It actually does help with readability and that's even more true now
that you have to take into account the location information on hir values.

This PR adds a couple of methods like this to the `hir::Value` type.
It's quite similar to `serde_json::Value`.