breaking(all): rename `ast` to `cst`
The Rowan-based typed syntax tree emitted by the parser used to be called
Abstract Syntax Tree (AST) but is in fact not very abstract: it preserves
text input losslessly, and all tree leaves are string-based tokens.
This renames it to Concrete Syntax Tree (CST)
and renames various APIs accordingly.
This leaves the AST name available for a new tree representation (coming soon).