Validate input value types
This is a pretty big thing that we missed until now, so I added it to
the current validation implementation.

Input values (in `input object` declarations and in argument
declarations) did not have full validation. Only their names were
checked, but if you declared a `arg: NonExistentType`, it would happily
truck on.

This adds a check, similar to object type field validation's output type
check. If you declare an input value or argument with an output type,
you get an error, if you provide a type that doesn't exist, you get an
error.