Revert from FileId::with_deterministic_ids back to serial_test
The approach in 41071dd524486cb2e36712c3cd36673ac037ea5f was flawed:
the mutex would only be effective if all other code running concurrently
also used it. As this wasn’t the case, there occasionally was some
flaky test failures.

This reverts to using the serial_test crate, and additionally moves
revelant tests to a separate executable so that all tests in that process
run serially.