Make everything know their own name
Fixes https://github.com/apollographql/apollo-rs/issues/708

In a few places (but not consistently) a `name` field
was omitted from some structs used as map values
on the basis that it would have been redundant with the map key.
This reverts that decision,
making it the user’s responsibility when mutating documents to keep names consistent.

* Add a `pub name: Name` field to `executable::Fragment` as well as
  `ScalarType`, `ObjectType`, `InterfaceType`, `EnumType`, `UnionType`, and `InputObjectType`
  in `schema`.
* Add a `fn name(&self) -> &Name` method to the `schema::ExtendedType` enum
* Add a `pub name: Option<Name>` field to `executable::Operation`
* Remove `executable::OperationRef<'_>`
  (which was equivalent to `(Option<&Name>, &Node<Operation>)`),
  replacing its uses with `&Node<Operation>`