Fix serializing single-line strings with leading whitespace
When parsing a block string, the first line does not affect the common
indent, as it would normally have no indent at all. But when
serializing a string to a block string, the first line *is* important,
namely for a single-line string that starts with whitespace.

Assume a text like this:
```
    Starts with "spaces"
```

If we skip the first line, the common indent is detected as 0,
producing this output:
```graphql
"""
    Starts with "spaces"
"""
```

Now when reparsing, the common indent is detected as 4, and you get
back:
```
Starts with "spaces"
```
which is wrong.