fix(lexer): lex \\""" as \ + \""" in block strings
Backslash is not special in block strings except in the exact sequence
`\"""`. So if you write `\\"""`, that sequence should still be treated
as the escape sequence for `"""`. This patch fixes that.

Previously, we would return to the BlockStringLiteral state on the
second `\`, and interpret the `"""` as the end of the string.