fix(parser,compiler): fix string serialization bugs found by fuzzing (#774)
* fuzz: add a fuzz target for string encoding/parsing

* Update to use compiler serialization

* Fix serializing single-line strings with leading whitespace

When parsing a block string, the first line does not affect the common
indent, as it would normally have no indent at all. But when
serializing a string to a block string, the first line *is* important,
namely for a single-line string that starts with whitespace.

Assume a text like this:
```
    Starts with "spaces"
```

If we skip the first line, the common indent is detected as 0,
producing this output:
```graphql
"""
    Starts with "spaces"
"""
```

Now when reparsing, the common indent is detected as 4, and you get
back:
```
Starts with "spaces"
```
which is wrong.

* Add failing test for \\""" in string

* enable debug log in fuzz test

* fix(lexer): lex \\""" as \ + \""" in block strings

Backslash is not special in block strings except in the exact sequence
`\"""`. So if you write `\\"""`, that sequence should still be treated
as the escape sequence for `"""`. This patch fixes that.

Previously, we would return to the BlockStringLiteral state on the
second `\`, and interpret the `"""` as the end of the string.

* fix rebase mistake

* changelogs

* pr number