chore(parser): reduce intermediate allocations
By threading the appropriate input lifetime through lexer and parser
methods, we can avoid copying strings in a few places.

Especially nice is that we can avoid copying the same strings multiple
times when peeking at tokens, and we can avoid an eager copy that is
almost always just thrown away in the implementation of `.expect()`.