Reduce duplicate validation errors from field merging
Field merging validation was applied to operations and fragment
definitions. Selection sets of fragment definitions must always be
checked in the context of an operation or other fragment, and in the end
it always leads to an operation. So we strictly only need to validate
operations to find all issues in all reachable fragments.

This can still output duplicate errors if a fragment contains a conflict
internally, and is used in multiple operations.