fix(compiler): cache the implementers map per operation
Fixes #862.

I had proposed a different approach that would let us cache the
implementers map for as long as the schema is immutable, so it could be
reused for different query validations. That approach had an issue that
made `Valid::assume_valid_ref` very, very subtle to use, so we are not
doing that right now.

This is a less ideal version but it does solve the immediate problem.
We already pass around an `OperationValidationConfig` structure and
this seems like a nice, non-invasive place to put the implementers
cache.