fix(compiler): remove potential hash collision bug from FieldsInSetCanMerge
To avoid cloning a big `Vec` for use as a hash key, I instead used the
hash of the vec as a key. I guess it's possible to get a hash collision
that way and produce the wrong result.

Here I avoid the clone by allocating field selections into a bump
allocator. The bump allocator outlives `FieldsInSetCanMerge` so we can
pass around references to slices. We already didn't deallocate the
`Vec`s so that should be the same.

The only issue I see here is that we build some simultaneously which
will behave relatively poorly with a bump allocator.