Introduce a faster hasher
This changeset introduces ahash, a hasher that is faster than the one provided by the stdlib.

It exposes a collections::fast module, that can be used by other libraries that rely on apollo-rs types.
This makes it easier for the apollo-rs to swap implementations in the future if needs be.