chore(compiler): make unused variable validation more efficient
The validation used a hash map and two hash sets. We can do the exact
same validation with only a hash map, by first putting all declared variables
in the hash map, and removing variables as we see them used. Then at the end we
only have unused variables in the map.