fix(compiler): fix validation for fragment spread of interface without implementers (#896)
* fix(compiler): add failing test for spread of empty interface

`Intf` has no implementations. As written in the spec, doing a `... on
Intf` fragment spread should never work, as the set of possible types is
empty and can never intersect with the parent type. However,
implementations like graphql-js and graphql-go have an early check,
accepting the fragment if the type condition is equal to the parent
type. This tests reproduces that.

We may want to align with graphql-js and graphql-go rather than the spec
here for compatibility? Though it's not something that's likely to
happen in the real world.

Ref https://github.com/graphql/graphql-spec/issues/1109

* fix(compiler): always accept spreading ...Ty when parent type is Ty