Validation adds/removes built-in scalar defs based on usage
We choose to reflect in the Rust API the behavior expected of introspection:

https://spec.graphql.org/draft/#sec-Scalars.Built-in-Scalars
> When returning the set of types from the `__Schema` introspection type,
> all referenced built-in scalars must be included.
> If a built-in scalar type is not referenced anywhere in a schema
> (there is no field, argument, or input field of that type)
> then it must not be included.

Because our public API allows arbitrary mutation of a Rust `Schema`,
we need a full schema scan at some point to determine usage.
Validation is a good fit as it already does this scan.

In [the full GraphQL grammar](https://spec.graphql.org/draft/#sec-Document-Syntax)
the `NamedType` rule is used in:

* `TypeCondition`: not in a schema
* `RootOperationDefinition`, `UnionMemberTypes`: must refer to an object type
* `ImplementsInterfaces`: must refer to an interface type, not a scalar
* `Type` / `NonNullType`:
  * … in `VariableDefinition`: not in a schema
  * … in `FieldDefinition`: object/interface field "return" type, **can be a scalar**
  * … in `InputValueDefinition`: **can be a scalar**
    * … in `InputFieldsDefinition` (fields of input types)
    * … in `ArgumentsDefinition` for directive arguments
      or object/interface field arguments

Based on this, I think `validate_input_value_definitions` and
`validate_field_definitions` are the only two places where validation code
needs to record potential references to built-in scalar types.