Silence clippy::result_large_err warning (#922)
This lint seems to be more aggressive in Rust 1.82 than in previous versions
as it now warns about functions returning `Result<Schema, WithErrors<Schema>>`
or `Result<ExecutableDocument, WithErrors<ExecutableDocument>>`.

The lint is correct that a return value of e.g. 224 bytes has some cost,
but parsing and validation functions are typically not called very often
so we accept this cost rather than make the API more awkward
by boxing schemas and documents.