Make variables in const contexts validation errors
Fixes https://github.com/apollographql/apollo-rs/issues/852

GraphQL values can be of different kinds, including variables with the `$x` syntax.
Variables are not allowed in some contexts, such as schemas.
The parser already rejects them as syntax errors, but the Rust API allows mutating
a document and representing variable values in contexts where they are not allowed.

It was already the case that would usually cause a validation error like
``variable `$x` is not defined``. This PR ensures and tests that this is the case
for every relevant context. It also makes related drive-by fixes:

* Validate the default value of input fields and arguments
* Validate the default value of variables once at their definition,
  not at every usage of the variable. This fixes duplicate diagnostics
  for a default value of incorrect type for a variable correctly used multiple times.