fix(smith): bump minimum `arbitrary` version (#1007)
apollo-smith uses the `.choose_iter()` method, which was introduced in
arbitrary v1.4.0. If used in a project that already depends on an older
version of arbitrary, you could have problems.

I ran into this when I pulled the latest changes into a very old clone
of the repo and ran tests: I still had 1.3.0 installed so compilation
failed.